+++ plone/autoform/form.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from plone.autoform.base import AutoFields
 from plone.autoform.interfaces import IAutoExtensibleForm
 from plone.autoform.interfaces import IAutoObjectSubForm
@@ -20,19 +19,18 @@ class AutoExtensibleForm(AutoFields, ExtensibleForm):
     @property
     def schema(self):
         raise NotImplementedError(
-            'The class deriving from AutoExtensibleForm must have a '
-            '\'schema\' property'
+            "The class deriving from AutoExtensibleForm must have a "
+            "'schema' property"
         )
 
     @property
     def additionalSchemata(self):
-        """Default to there being no additional schemata
-        """
+        """Default to there being no additional schemata"""
         return ()
 
     def updateFields(self):
         self.updateFieldsFromSchemata()
-        super(AutoExtensibleForm, self).updateFields()
+        super().updateFields()
 
 
 @implementer(IAutoObjectSubForm)