+++ plone/autoform/widgets.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from plone.autoform.interfaces import IParameterizedWidget
 from plone.autoform.interfaces import IWidgetExportImportHandler
 from plone.autoform.utils import resolveDottedName
@@ -17,12 +16,11 @@ from zope.interface import implementer
 from zope.interface import providedBy
 from zope.schema import getFields
 
-import six
 import z3c.form.browser.interfaces
 
 
 @implementer(IParameterizedWidget)
-class ParameterizedWidget(object):
+class ParameterizedWidget:
     """A factory for deferred construction of widgets with parameters.
 
     z3c.form widgets are associated with a particular request,
@@ -44,49 +42,51 @@ class ParameterizedWidget(object):
 
     def __init__(self, widget_factory=None, **params):
         if widget_factory is not None:
-            if not IFieldWidget.implementedBy(widget_factory) \
-                    and not IWidget.implementedBy(widget_factory) \
-                    and not isinstance(widget_factory, six.string_types):
-                raise TypeError('widget_factory must be an IFieldWidget '
-                                'or an IWidget')
+            if (
+                not IFieldWidget.implementedBy(widget_factory)
+                and not IWidget.implementedBy(widget_factory)
+                and not isinstance(widget_factory, str)
+            ):
+                raise TypeError(
+                    "widget_factory must be an IFieldWidget " "or an IWidget"
+                )
         self.widget_factory = widget_factory
         self.params = params
 
     def __call__(self, field, request):
         __traceback_info__ = (
-            '{0}, processing:\n'
-            '- field "{1}"\n'
-            '- widget: {2}\n'
-            '- params: {3}\n'.format(
+            "{}, processing:\n"
+            '- field "{}"\n'
+            "- widget: {}\n"
+            "- params: {}\n".format(
                 self.__class__.__name__,
                 field.__name__,
                 repr(self.widget_factory),
-                self.params
+                self.params,
             )
         )
-        if isinstance(self.widget_factory, six.string_types):
-            __traceback_info__ += '- resolving dotted name\n'
+        if isinstance(self.widget_factory, str):
+            __traceback_info__ += "- resolving dotted name\n"
             self.widget_factory = resolveDottedName(self.widget_factory)
         if self.widget_factory is None:
             # use default widget factory for this field type
-            __traceback_info__ += '- using default widget factory\n'
+            __traceback_info__ += "- using default widget factory\n"
             widget = getMultiAdapter((field, request), IFieldWidget)
         elif IWidget.implementedBy(self.widget_factory):
-            __traceback_info__ += '- calling factory, then wrapping with ' \
-                                  'FieldWidget\n'
+            __traceback_info__ += (
+                "- calling factory, then wrapping with " "FieldWidget\n"
+            )
             widget = FieldWidget(field, self.widget_factory(request))
         elif IFieldWidget.implementedBy(self.widget_factory):
-            __traceback_info__ += '- calling factory\n'
+            __traceback_info__ += "- calling factory\n"
             widget = self.widget_factory(field, request)
         for k, v in self.params.items():
             setattr(widget, k, v)
         return widget
 
     def __repr__(self):
-        return '{0}({1}, {2})'.format(
-            self.__class__.__name__,
-            self.widget_factory,
-            self.params
+        return "{}({}, {})".format(
+            self.__class__.__name__, self.widget_factory, self.params
         )
 
     def getWidgetFactoryName(self):
@@ -97,8 +97,8 @@ class ParameterizedWidget(object):
         widget = self.widget_factory
         if widget is None:
             return
-        if not isinstance(widget, six.string_types):
-            widget = '{0}.{1}'.format(widget.__module__, widget.__name__)
+        if not isinstance(widget, str):
+            widget = f"{widget.__module__}.{widget.__name__}"
         return widget
 
     def getExportImportHandler(self, field):
@@ -110,25 +110,23 @@ class ParameterizedWidget(object):
             # instantiate the widget.
             sm = getSiteManager()
             widgetFactory = sm.adapters.lookup(
-                (providedBy(field), IFormLayer), IFieldWidget)
+                (providedBy(field), IFormLayer), IFieldWidget
+            )
             if widgetFactory is not None:
-                widgetName = '{0}.{1}'.format(
-                    widgetFactory.__module__,
-                    widgetFactory.__name__
+                widgetName = "{}.{}".format(
+                    widgetFactory.__module__, widgetFactory.__name__
                 )
             else:
-                widgetName = u''
+                widgetName = ""
 
-        widgetHandler = queryUtility(IWidgetExportImportHandler,
-                                     name=widgetName)
+        widgetHandler = queryUtility(IWidgetExportImportHandler, name=widgetName)
         if widgetHandler is None:
             widgetHandler = WidgetExportImportHandler(IHTMLFormElement)
         return widgetHandler
 
 
 @implementer(IWidgetExportImportHandler)
-class WidgetExportImportHandler(object):
-
+class WidgetExportImportHandler:
     def __init__(self, widget_schema):
         self.fieldAttributes = getFields(widget_schema)
 
@@ -136,10 +134,7 @@ class WidgetExportImportHandler(object):
         for attributeName, attributeField in self.fieldAttributes.items():
             for node in widgetNode.iterchildren():
                 if noNS(node.tag) == attributeName:
-                    params[attributeName] = elementToValue(
-                        attributeField,
-                        node
-                    )
+                    params[attributeName] = elementToValue(attributeField, node)
 
     def write(self, widgetNode, params):
         for attributeName, attributeField in self.fieldAttributes.items():