+++ CHANGES.rst
@@ -150,14 +150,14 @@ Incompatibilities:
 
 New:
 
-- Fieldset labels/descriptions we're taken from first occurence.
+- Fieldset labels/descriptions we're taken from first occurrence.
   It was not possible to override them in a subsequent directive.
   Also it was not possible to set them in a subsequent directive, if it was not set before.
   Now subsequent directives w/o a label/description are just adding the field to the fieldset.
   If a different label and/or description is given, it replaces the existing prior loaded one.
   [jensens]
 
-- The order of the fieldsets can be defined now explictly with the ``plone.supermodel.directives.fieldset`` directive.
+- The order of the fieldsets can be defined now explicitly with the ``plone.supermodel.directives.fieldset`` directive.
   ``plone.autoform`` now does the sorting while fieldset processing.
   [jensens]